+++ prince-archiver/README.md
@@ -5,24 +5,24 @@ object storage backend as soon as a timestep is available.
 
 It consists of the following entrypoints:
 - `event-ingester`
-- `worker`
+- `state-manager`
 - `upload-worker`
 - `api`
 
-The `data-ingester` ingests imaging events produced by various imaging systems (e.g. 
+The `event-ingester` ingests imaging events produced by various imaging systems (e.g. 
 Prince), copying them to a staging area if necessary. It then appends this information 
 to a redis stream which can be read by other (sub)systems. It is also responsible for 
 deleting source and staging files after a predefined amount of elapsed time.
 
-The `worker` is responsible for persisting state relating to imaging events, 
-storing it in a local database. It pulls messages from the `data-ingester` stream and 
+The `state-manaher` is responsible for persisting state relating to imaging events, 
+storing it in a local database. It pulls messages from the `event-ingester` stream and 
 also recieves notifications from the `upload-worker` and external `surf-archiver` systems.
 The former notifies the worker that image data has been uploaded to object storage. The 
 latter informs the system when uploads have been moved into the tape archive.
 
 The `upload-worker` pulls messages from the `data-ingester` stream. It bundles the 
 corresponding image data into a tar file, and uploads it to an S3 supported object
-storage. It notifies the `worker` when this upload process is complete.
+storage. It notifies the `state-manager` when this upload process is complete.
 
 The `api` acts as a gateway to access information related to the latest state of the 
 imaging events.
@@ -31,7 +31,7 @@ imaging events.
 # Input directory structure 
 
 In order to successfully ingest imaging events, the file system of the external systems 
-(e.g. Prince, Aretha) need to be mounted into the the `data-ingester` container. 
+(e.g. Prince, Aretha) need to be mounted into the `event-ingester` container. 
 It is assumed that the directory structure of the source systems has the following 
 stucture. The data ingestion
